+++ src/Cirreum.Invocation.WebSockets/WebSocketInvocationRegistrar.cs
@@ -4,7 +4,6 @@ using Cirreum.Invocation.Configuration;
 using Cirreum.Invocation.WebSockets;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;